requirements.txt
gradio
langchain
langchain_core
langchain_openai
nltk
openai
pandas
PyYAML
